refactor: extract `build_ic_map`
